+++ src/jsandbox.js
@@ -1,5 +1,5 @@
 /*
- * JSandbox JavaScript Library v0.2.1
+ * JSandbox JavaScript Library v0.2.2
  * 2009-11-08
  * By Elijah Grey, http://eligrey.com
  *
@@ -7,20 +7,26 @@
  *   See COPYING.md
  */
 
-/*global Worker, JSON*/
+/*global self */
 
-/*jslint white: true, onevar: true, browser: true, undef: true, nomen: true, eqeqeq: true,
-bitwise: true, regexp: true, newcap: true, immed: true, maxerr: 1000, maxlen: 90 */
+/*jslint white: true, undef: true, nomen: true, eqeqeq: true, bitwise: true, regexp: true,
+newcap: true, immed: true, maxerr: 1000, maxlen: 90 */
 
-var Sandbox = (function () {
+var JSandbox = (function (self, globalEval) {
 	"use strict";
 	
-	if (typeof Worker === "undefined") {
+	var undef_type = "undefined",
+	doc = self.document,
+	Worker = self.Worker;
+	
+	if (typeof Worker === undef_type || typeof doc === undef_type) {
 		return;
 	}
 	
 	var
-	// repeatedly used strings (for minification)
+	JSON = self.JSON,
+	
+	// repeatedly used properties/strings (for minification)
 	$eval       = "eval",
 	$exec       = "exec",
 	$load       = "load",
@@ -33,9 +39,10 @@ var Sandbox = (function () {
 	$worker     = "worker",
 	$onresponse = "onresponse",
 	$prototype  = "prototype",
+	$call       = "call",
 	
-	$str_type   = "string",
-	$fun_type   = "function",
+	str_type   = "string",
+	fun_type   = "function",
 	
 	
 	Sandbox = function () {
@@ -50,7 +57,7 @@ var Sandbox = (function () {
 		
 		sandbox[$worker].onmessage = function (event) {
 			var data = event[$data], request;
-			if (typeof data === $str_type) { // parse JSON
+			if (typeof data === str_type) { // parse JSON
 				try {
 					data = JSON.parse(data);
 				} catch (e) {
@@ -63,19 +70,19 @@ var Sandbox = (function () {
 			request = sandbox[$requests][data.id];
 			if (request) {
 				if (data.error) {
-					if (typeof sandbox[$onerror] === $fun_type) {
+					if (typeof sandbox[$onerror] === fun_type) {
 						sandbox[$onerror](data, request);
 					}
-					if (typeof request[$onerror] === $fun_type) {
-						request[$onerror].call(sandbox, data.error);
+					if (typeof request[$onerror] === fun_type) {
+						request[$onerror][$call](sandbox, data.error);
 					}
 				} else {
-					if (typeof sandbox[$onresponse] === $fun_type) {
+					if (typeof sandbox[$onresponse] === fun_type) {
 						sandbox[$onresponse](data, request);
 					}
 				
-					if (typeof request[$callback] === $fun_type) {
-						request[$callback].call(sandbox, data.results);
+					if (typeof request[$callback] === fun_type) {
+						request[$callback][$call](sandbox, data.results);
 					}
 				}
 				delete sandbox[$requests][data.id];
@@ -83,10 +90,13 @@ var Sandbox = (function () {
 		};
 	},
 	proto = Sandbox[$prototype],
+	jsonStringifyThis = function () { // for compatability with WebKit/Chrome
+		return JSON.stringify(this);
+	},
 	createRequestMethod = function (method) {
 		proto[method] = function (options, callback, input, onerror) {
-			if (typeof options === $str_type ||
-			    Object[$prototype].toString.call(options) === "[object Array]" ||
+			if (typeof options === str_type ||
+			    Object[$prototype].toString[$call](options) === "[object Array]" ||
 			    arguments.length > 1)
 			{ // called in (data, callback, input, onerror) style
 				options = {
@@ -97,7 +107,7 @@ var Sandbox = (function () {
 				};
 			}
 			
-			if (method === $load && typeof options[$data] === $str_type) {
+			if (method === $load && typeof options[$data] === str_type) {
 				options[$data] = [options[$data]];
 			}
 			
@@ -116,9 +126,7 @@ var Sandbox = (function () {
 				method   : method,
 				data     : data,
 				input    : input,
-				toString : function () { // for compatability with WebKit/Chrome
-					return JSON.stringify(this);
-				}
+				toString : jsonStringifyThis
 			});
 		
 			return id;
@@ -133,16 +141,26 @@ var Sandbox = (function () {
 		
 			Sandbox[$prototype][method].apply(
 				sandbox,
-				Array[$prototype].slice.call(arguments)
+				Array[$prototype].slice[$call](arguments)
 			);
 			return Sandbox;
 		};
 	},
-	doc = document,
 	linkElems = doc.getElementsByTagName("link"),
-	linkElem,
 	methods = [$eval, $load, $exec],
-	i = 3;
+	i = 3, // methods.length
+	ready = true,
+	readyQueue = [];
+	
+	Sandbox.ready = function (fn) {
+		if (ready) {
+			try {
+				fn();
+			} catch (e) {}
+		} else {
+			readyQueue.unshift(fn);
+		}
+	};
 	
 	while (i--) {
 		createRequestMethod(methods[i]);
@@ -165,33 +183,36 @@ var Sandbox = (function () {
 		return id;
 	};
 	
-	(typeof JSON === "undefined" &&
-		((doc.getElementsByTagName("head")[0] || doc.documentElement)
-			.appendChild(doc.createElement("script")).src =
-				"http://github.com/eligrey/jsandbox/raw/master/min/json2.js"
-		)
-	);
+	if (typeof JSON === undef_type) {
+		ready = false;
+		var jsonScript = doc.createElement("script");
+		jsonScript.type = "text/javascript";
+		jsonScript.onload = function () {
+			if (self.JSON) {
+				JSON = self.JSON;
+				ready = true;
+				var i = readyQueue.length;
+			
+				while (i--) {
+					try {
+						readyQueue.pop()();
+					} catch (e) {}
+				}
+			}
+		};
+		jsonScript.src = "http://github.com/eligrey/jsandbox/raw/master/min/json2.js";
+		doc.documentElement.appendChild(jsonScript);
+	}
 	
 	i = linkElems.length;
 	while (i--) {
-		linkElem = linkElems[i];
-		if (linkElem.getAttribute("rel") === "jsandbox" &&
-		    linkElem.hasAttribute("href"))
+		if (linkElems[i].getAttribute("rel") === "jsandbox")
 		{
-			Sandbox.url = linkElem.getAttribute("href");
+			Sandbox.url = linkElems[i].getAttribute("href");
 			break;
 		}
 	}
 	
-	// dereference no-longer used variables
-	proto               =
-	createRequestMethod =
-	doc                 =
-	linkElems           =
-	linkElem            =
-	methods             =
-	i                   =
-	undefined;
-	
 	return Sandbox;
-}());
+}(self, eval)),
+Sandbox = JSandbox;